medicine add

+++ app/src/main/java/com/dreamdigitizers/drugmanagement/presenters/implementations/PresenterCapturedPicturePreview.java
@@ -1,5 +1,8 @@
 package com.dreamdigitizers.drugmanagement.presenters.implementations;
 
+import android.graphics.Bitmap;
+import android.text.TextUtils;
+
 import com.dreamdigitizers.drugmanagement.presenters.abstracts.IPresenterCapturedPicturePreview;
 import com.dreamdigitizers.drugmanagement.utils.FileUtils;
 import com.dreamdigitizers.drugmanagement.views.abstracts.IViewCapturedPicturePreview;
@@ -11,7 +14,15 @@ class PresenterCapturedPicturePreview implements IPresenterCapturedPicturePrevie
         this.mView = pView;
     }
 
-    public void deleteFile(String pFilePath) {
-        FileUtils.deleteFile(pFilePath);
+    @Override
+    public Bitmap loadImage(String pFilePath, int pWidth, int pHeight) {
+        return FileUtils.decodeSampledBitmapFromFile(pFilePath, pWidth, pHeight);
+    }
+
+    @Override
+    public void deleteImage(String pFilePath) {
+        if(!TextUtils.isEmpty(pFilePath)) {
+            FileUtils.deleteFile(pFilePath);
+        }
     }
 }
